#80 feat : TdsGraphContent
