+++ feature/log/src/main/kotlin/com/titi/app/feature/log/ui/DailyScreen.kt
@@ -5,17 +5,11 @@ import android.graphics.Picture
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
@@ -24,7 +18,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
@@ -33,11 +26,7 @@ import androidx.compose.ui.unit.dp
 import com.titi.app.core.designsystem.R
 import com.titi.app.core.designsystem.component.TdsColorRow
 import com.titi.app.core.designsystem.component.TdsDialog
-import com.titi.app.core.designsystem.component.TdsPagerIndicator
-import com.titi.app.core.designsystem.component.TdsStandardDailyGraph
-import com.titi.app.core.designsystem.component.TdsTaskProgressDailyGraph
-import com.titi.app.core.designsystem.component.TdsTimeLineDailyGraph
-import com.titi.app.core.designsystem.component.TdsTimeTableDailyGraph
+import com.titi.app.core.designsystem.component.TdsGraphContent
 import com.titi.app.core.designsystem.model.TdsDialogInfo
 import com.titi.app.core.designsystem.model.TdsTaskData
 import com.titi.app.core.designsystem.model.TdsTimeTableData
@@ -167,7 +156,7 @@ fun DailyScreen(
 
         Spacer(modifier = Modifier.height(15.dp))
 
-        GraphContent(
+        TdsGraphContent(
             modifier = Modifier.fillMaxWidth(),
             todayDate = currentDate.toString().replace('-', '.'),
             todayDayOfTheWeek = currentDate.dayOfWeek.value - 1,
@@ -180,110 +169,6 @@ fun DailyScreen(
             checkedButtonStates = checkedButtonStates,
             pictureList = pictureList,
             onCheckedChange = onCheckedChange,
-        )
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun GraphContent(
-    modifier: Modifier = Modifier,
-    todayDate: String,
-    todayDayOfTheWeek: Int,
-    totalTime: String,
-    maxTime: String,
-    taskData: List<TdsTaskData>,
-    tdsColors: List<TdsColor>,
-    timeLines: List<Long>,
-    timeTableData: List<TdsTimeTableData>,
-    checkedButtonStates: List<Boolean>,
-    pictureList: List<Picture>,
-    onCheckedChange: (page: Int, checked: Boolean) -> Unit,
-) {
-    val pagerState = rememberPagerState(
-        pageCount = {
-            4
-        },
-    )
-
-    Box(
-        modifier = modifier,
-        contentAlignment = Alignment.BottomCenter,
-    ) {
-        HorizontalPager(
-            modifier = Modifier.wrapContentSize(),
-            userScrollEnabled = true,
-            state = pagerState,
-            beyondBoundsPageCount = 2,
-        ) { page ->
-            when (page % 4) {
-                0 -> TdsStandardDailyGraph(
-                    modifier = Modifier.fillMaxWidth(),
-                    todayDate = todayDate,
-                    todayDayOfTheWeek = todayDayOfTheWeek,
-                    tdsColors = tdsColors,
-                    timeLines = timeLines,
-                    taskData = taskData,
-                    totalTime = totalTime,
-                    maxTime = maxTime,
-                    picture = pictureList[0],
-                    checked = checkedButtonStates[0],
-                    onCheckedChange = {
-                        onCheckedChange(0, it)
-                    },
-                )
-
-                1 -> TdsTimeTableDailyGraph(
-                    modifier = Modifier.fillMaxWidth(),
-                    todayDate = todayDate,
-                    todayDayOfTheWeek = todayDayOfTheWeek,
-                    tdsColors = tdsColors,
-                    taskData = taskData,
-                    timeTableData = timeTableData,
-                    totalTime = totalTime,
-                    maxTime = maxTime,
-                    picture = pictureList[1],
-                    checked = checkedButtonStates[1],
-                    onCheckedChange = {
-                        onCheckedChange(1, it)
-                    },
-                )
-
-                2 -> TdsTimeLineDailyGraph(
-                    modifier = Modifier.fillMaxWidth(),
-                    todayDate = todayDate,
-                    todayDayOfTheWeek = todayDayOfTheWeek,
-                    tdsColors = tdsColors,
-                    timeLines = timeLines,
-                    totalTime = totalTime,
-                    maxTime = maxTime,
-                    picture = pictureList[2],
-                    checked = checkedButtonStates[2],
-                    onCheckedChange = {
-                        onCheckedChange(2, it)
-                    },
-                )
-
-                3 -> TdsTaskProgressDailyGraph(
-                    modifier = Modifier.fillMaxWidth(),
-                    todayDate = todayDate,
-                    taskData = taskData,
-                    tdsColors = tdsColors,
-                    picture = pictureList[3],
-                    checked = checkedButtonStates[3],
-                    onCheckedChange = {
-                        onCheckedChange(3, it)
-                    },
-                )
-            }
-        }
-
-        TdsPagerIndicator(
-            modifier = Modifier.padding(bottom = 8.dp),
-            pagerState = pagerState,
-            indicatorCount = 4,
-            indicatorSize = 8.dp,
-            activeColor = TdsColor.TEXT.getColor(),
         )
     }
 }
